+++ Stumpy/POP3/POPSessionHandler.swift
@@ -14,7 +14,8 @@ final class POPSessionHandler: ChannelInboundHandler {
 
     let sessionState: POPSessionState
 
-    init(with store: MailStore, hostName: String) {
+    init(with store: MailStore,
+         hostName: String) {
         sessionState = POPSessionState(with: store, hostName: hostName)
     }
 
@@ -39,11 +40,12 @@ final class POPSessionHandler: ChannelInboundHandler {
         // then we'd care more about this. The one thing we should care about is that this string
         // is different for each request.
         // swiftlint:disable:next line_length
-        let message = "Stumpy POP3 server ready <\(sessionState.sessionID).\(now.timeIntervalSinceReferenceDate)@\(sessionState.hostName)>\r\n"
+        let message = "+OK Stumpy POP3 server ready <\(sessionState.sessionID).\(now.timeIntervalSinceReferenceDate)@\(sessionState.hostName)>\n"
         var outBuff = context.channel.allocator.buffer(capacity: message.count)
         outBuff.writeString(message)
 
         context.writeAndFlush(NIOAny(outBuff), promise: nil)
         sessionState.popState = .authorization
+        context.fireChannelActive()
     }
 }