+++ Stumpy/SMTP/SMTPActionHandler.swift
@@ -13,6 +13,7 @@ import NIO
 // what the incoming action is, and how we should mutate the state
 // and what message we should send back to the client.
 
+// swiftlint:disable:next type_body_length
 final class SMTPActionHandler: ChannelInboundHandler {
     typealias InboundIn = SMTPSessionState
     typealias InboundOut = ByteBuffer
@@ -21,51 +22,116 @@ final class SMTPActionHandler: ChannelInboundHandler {
 
     init() {
         logger = Logger(label: "SMTPActionHandler")
+        logger.logLevel = .info
         logger[metadataKey: "origin"] = "[SMTP]"
     }
 
     public func channelRead(context: ChannelHandlerContext, data: NIOAny) {
         let currentState = unwrapInboundIn(data)
+        var gotTermination = false
+
+        // now, we look at the line that came in and turn it into an action
+        if currentState.smtpState != .dataBody && currentState.smtpState != .dataHeader {
+            processMessage(currentState)
+        } else {
+            // we need to buffer all the incoming data until we get the termination sequence.
+            // Once we get that, *then* we can process the message data.
+            currentState.accumulatedData.append(currentState.inputLine)
+            // check for termination
+            gotTermination = currentState.accumulatedData
+                .data(using: .utf8)?.lines().contains(where: { $0 == "." }) ?? false
+        }
 
         // Figuring out what comes next might take some time. We're going
         // to go async and do that work, and the rest of the pipeline
         // will just wait for us.
-        Task {
-            let response = await computeResponse(currentState)
-            // now, we *might* want to mutate the current message
-            store(currentState)
-            // and we *might* want to put the message into the mail store
-            if currentState.smtpState == .quit {
-                if hasMessageIDHeader(currentState.workingMessage) {
-                    let saveMessage = currentState.workingMessage
-                    currentState.workingMessage = MemoryMessage()
-                    Task {
-                        await currentState.mailstore.add(message: saveMessage)
-                    }
-                } else {
-                    logger.info("No message-id header, so not saving the working message")
-                }
-            }
+        if gotTermination || (currentState.smtpState != .dataHeader && currentState.smtpState != .dataBody) {
+            spinTask(context: context, currentState: currentState)
+        }
+    }
 
-            // and finally, we send a response message back to the client
-            var respMessage = ""
-            if response.code > 0 {
-                // we send back the numeric code -- unless we're just accepting
-                // data for a mail message
-                respMessage.append("\(response.code) \(response.message)")
-            }
-            if !respMessage.hasSuffix("\r\n") {
-                respMessage.append("\r\n")
-            }
-            context.eventLoop.execute {
-                var outBuffer = context.channel.allocator.buffer(capacity: respMessage.count)
-                outBuffer.writeString(respMessage)
-                context.writeAndFlush(self.wrapInboundOut(outBuffer), promise: nil)
-                if currentState.smtpState == .quit {
-                    _ = context.close()
+    private func spinTask(context: ChannelHandlerContext, currentState: SMTPSessionState) {
+        Task {
+            var response: SMTPResponse = .badSequence
+            if currentState.smtpState == .dataHeader || currentState.smtpState == .dataBody {
+                // this could be a multi-line sequence. The whole message, in fact.
+                // SO we need to split it up and process line by line.
+                let lines = currentState.accumulatedData.data(using: .utf8)?.lines() ?? []
+                if lines.isEmpty {
+                    logger.trace("Only one line incoming", metadata: ["line": "\(currentState.inputLine)"])
+                    response = await doAThing(currentState: currentState)
+                } else {
+                    for line in lines {
+                        logger.trace("Processing line", metadata: ["line": "\(line)"])
+                        currentState.inputLine = String(line)
+                        if currentState.smtpState == .dataHeader {
+                            logger.trace("It's a header")
+                            processDataHeaderMessage(currentState)
+                        } else if currentState.smtpState == .dataBody {
+                            logger.trace("It's a body line")
+                            processDataBodyMessage(currentState)
+                        }
+                        response = await doAThing(currentState: currentState)
+                        if currentState.smtpState == .quit || // session end
+                            currentState.smtpState == .mail || // done with the message
+                            response.code > 299 // error condition
+                        { break }
+                    }
                 }
+                // and finally, we send a response message back to the client
+                if currentState.smtpState != .dataBody &&
+                    currentState.smtpState != .dataHeader {
+                    sendResponse(context: context, currentState: currentState, response: response)
+                }
+            } else {
+                response = await doAThing(currentState: currentState)
+                sendResponse(context: context, currentState: currentState, response: response)
             }
         }
+    }
+
+    private func sendResponse(context: ChannelHandlerContext,
+                              currentState: SMTPSessionState,
+                              response: SMTPResponse) {
+        var respMessage = ""
+        if response.code > 0 {
+            // we send back the numeric code -- unless we're just accepting
+            // data for a mail message
+            respMessage.append("\(response.code) \(response.message)")
+        }
+        context.eventLoop.execute {
+            if !(respMessage.hasSuffix("\r\n") || respMessage.hasSuffix("\n") || respMessage.isEmpty) {
+                respMessage.append("\n")
+            }
+            var outBuffer = context.channel.allocator.buffer(capacity: respMessage.count)
+            outBuffer.writeString(respMessage)
+            context.writeAndFlush(self.wrapInboundOut(outBuffer), promise: nil)
+            if currentState.smtpState == .quit {
+                _ = context.close()
+            }
+        }
+    }
+
+    private func doAThing(currentState: SMTPSessionState) async -> SMTPResponse {
+        let response = await computeResponse(currentState)
+        // now, we *might* want to mutate the current message
+        store(currentState)
+        // and we *might* want to put the message into the mail store
+        if currentState.smtpState == .quit // end session
+            || currentState.smtpState == .mail // end of message
+        {
+            if hasMessageIDHeader(currentState.workingMessage) {
+                let saveMessage = currentState.workingMessage
+                currentState.workingMessage = MemoryMessage()
+                currentState.accumulatedData = ""
+                Task {
+                    await currentState.mailstore.add(message: saveMessage)
+                }
+            } else {
+                logger.info("No message-id header, so not saving the working message")
+            }
+        }
+        return response
     }
 
     /// The RFC states that there should be a message-id header, but it neglects to
@@ -108,7 +174,7 @@ final class SMTPActionHandler: ChannelInboundHandler {
                         state.workingMessage.appendHeader(value: value, to: state.lastHeader)
                     }
                 } else if state.smtpState == .dataBody {
-                    logger.trace("appending line to message")
+                    logger.trace("appending line to message", metadata: ["line": "\(input)"])
                     state.workingMessage.append(line: input)
                 } else {
                     logger.trace("SMTP state is: \(state.smtpState); no input stored to message")
@@ -139,6 +205,7 @@ final class SMTPActionHandler: ChannelInboundHandler {
 
             case .data:
                 if state.smtpState == .rcpt {
+                    state.smtpState = .dataHeader
                     return SMTPResponse(code: 354,
                                         message: "Start mail input; end with <CRLF>.<CRLF>")
                 } else {
@@ -147,9 +214,9 @@ final class SMTPActionHandler: ChannelInboundHandler {
 
             case .dataEnd:
                 if state.smtpState == .dataBody || state.smtpState == .dataHeader {
-                    state.smtpState = .quit
+                    state.smtpState = state.mailEndState
                     return SMTPResponse(code: 250,
-                                        message: "OK")
+                                        message: "OK, message accepted for delivery: queued as 1")
                 } else {
                     return .badSequence
                 }
@@ -167,13 +234,12 @@ final class SMTPActionHandler: ChannelInboundHandler {
                 if state.smtpState == .greet {
                     state.smtpState = .mail
                     return SMTPResponse(code: 250,
-                                        message: "local.stumpy Hello \(command.parameters)\r\n250 OK")
+                                        message: "local.stumpy Hello \(command.parameters)\n250 OK")
                 } else {
                     return .badSequence
                 }
 
             case .expn:
-                // TODO: probably, we should just come back with the same email that was provided.
                 return SMTPResponse(code: 252,
                                     message: "Not supported")
 
@@ -228,7 +294,6 @@ final class SMTPActionHandler: ChannelInboundHandler {
                 return SMTPResponse(code: 250, message: "OK")
 
             case .vrfy:
-                // TODO: again, it would be trivial to say sure, that's our guy
                 return SMTPResponse(code: 252, message: "Not Supported")
 
             case .unknown:
@@ -241,6 +306,83 @@ final class SMTPActionHandler: ChannelInboundHandler {
         } else {
             logger.critical("There's no command!!!")
             return .badSequence
+        }
+    }
+
+    // swiftlint:disable:next function_body_length
+    func processMessage(_ state: SMTPSessionState) {
+        let allCaps = state.inputLine.uppercased()
+        var params = state.inputLine
+        if allCaps.hasPrefix("EHLO ") {
+            if params.count > 5 {
+                params.removeFirst(5)
+            } else {
+                params = ""
+            }
+            state.command = SMTPCommand(action: .ehlo,
+                                        parameters: params.trimmingCharacters(in: .whitespacesAndNewlines))
+        } else if allCaps.hasPrefix("HELO") {
+            if params.count > 5 {
+                params.removeFirst(5)
+            } else {
+                params = ""
+            }
+            state.command = SMTPCommand(action: .helo,
+                                        parameters: params.trimmingCharacters(in: .whitespacesAndNewlines))
+        } else if allCaps.hasPrefix("MAIL FROM:") {
+            params.removeFirst(10)
+            state.command = SMTPCommand(action: .mail,
+                                        parameters: params.trimmingCharacters(in: .whitespacesAndNewlines))
+        } else if allCaps.hasPrefix("RCPT TO:") {
+            params.removeFirst(8)
+            state.command = SMTPCommand(action: .rcpt,
+                                        parameters: params.trimmingCharacters(in: .whitespacesAndNewlines))
+        } else if allCaps.hasPrefix("DATA") {
+            state.command = SMTPCommand(action: .data, parameters: "")
+        } else if allCaps.hasPrefix("QUIT") {
+            state.command = SMTPCommand(action: .quit, parameters: "")
+        } else if allCaps.hasPrefix("RSET") {
+            state.command = SMTPCommand(action: .rset, parameters: "")
+        } else if allCaps.hasPrefix("NOOP") {
+            state.command = SMTPCommand(action: .noop, parameters: "")
+        } else if allCaps.hasPrefix("EXPN") {
+            params.removeFirst(4)
+            state.command = SMTPCommand(action: .expn,
+                                        parameters: params.trimmingCharacters(in: .whitespacesAndNewlines))
+        } else if allCaps.hasPrefix("VRFY") {
+            params.removeFirst(4)
+            state.command = SMTPCommand(action: .vrfy,
+                                        parameters: params.trimmingCharacters(in: .whitespacesAndNewlines))
+        } else if allCaps.hasPrefix("HELP") {
+            state.command = SMTPCommand(action: .help, parameters: "")
+        } else if allCaps.hasPrefix("LIST") { // not actually an SMTP command; allows inspecting the mail store
+            params.removeFirst(4)
+            state.command = SMTPCommand(action: .list,
+                                        parameters: params.trimmingCharacters(in: .whitespacesAndNewlines))
+        } else {
+            state.command = SMTPCommand(action: .unknown,
+                                        parameters: params.trimmingCharacters(in: .whitespacesAndNewlines))
+        }
+    }
+
+    func processDataHeaderMessage(_ state: SMTPSessionState) {
+        if state.inputLine.trimmingCharacters(in: .whitespacesAndNewlines) == "." {
+            state.command = SMTPCommand(action: .dataEnd, parameters: "")
+        } else if state.inputLine.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty {
+            state.command = SMTPCommand(action: .blankLine, parameters: "")
+        } else {
+            state.command = SMTPCommand(action: .unknown,
+                                        parameters: state.inputLine.trimmingCharacters(in: .whitespacesAndNewlines))
+        }
+    }
+
+    func processDataBodyMessage(_ state: SMTPSessionState) {
+        if state.inputLine.trimmingCharacters(in: .whitespacesAndNewlines) == "." {
+            state.command = SMTPCommand(action: .dataEnd, parameters: "")
+        } else if state.inputLine.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty {
+            state.command = SMTPCommand(action: .unknown, parameters: "\n")
+        } else {
+            state.command = SMTPCommand(action: .unknown, parameters: state.inputLine)
         }
     }
 }